# Switch from protoc encoding to api compile encoding