--- a/gapii/cc/pack_encoder.h
+++ b/gapii/cc/pack_encoder.h
@@ -19,6 +19,7 @@
 
 #include <memory>
 #include <unordered_map>
+#include <utility>
 
 #include <google/protobuf/io/coded_stream.h>
 #include <google/protobuf/io/zero_copy_stream_impl_lite.h>
@@ -42,12 +43,31 @@ class PackEncoder {
 public:
     typedef std::shared_ptr<PackEncoder> SPtr;
 
+    typedef uint32_t TypeID;
+    typedef std::pair<TypeID, bool> TypeIDAndIsNew;
+
+    virtual ~PackEncoder() = default;
+
+    // type encodes the given type descriptor if it hasn't been already,
+    // returning the type identifier and a boolean indicating whether the type
+    // was encoded this call.
+    // type assumes the data pointer is stable between calls of the same type.
+    virtual TypeIDAndIsNew type(const char* name, size_t size, const void* data) = 0;
+
     // object encodes the leaf protobuf message.
     virtual void object(const ::google::protobuf::Message* msg) = 0;
+
+    // object encodes the leaf object from an already encoded protobuf message.
+    virtual void object(TypeID type, size_t size, const void* data) = 0;
 
     // group encodes the protobuf message as a group that can contain other
     // objects and groups.
     virtual SPtr group(const ::google::protobuf::Message* msg) = 0;
+
+    // object encodes the group object from an already encoded protobuf message.
+    // The returned PackEncoder can be used to encode objects into this group,
+    // and must be deleted by the caller.
+    virtual PackEncoder* group(TypeID type, size_t size, const void* data) = 0;
 
     // flush flushes out all of the pending in the encoder
     virtual void flush() = 0;
@@ -57,9 +77,6 @@ public:
 
     // noop returns a PackEncoder::SPtr that does nothing.
     static SPtr noop();
-
-protected:
-    ~PackEncoder() = default;
 };
 
 } // namespace gapii