--- a/gapis/api/gles/api/state_queries.api
+++ b/gapis/api/gles/api/state_queries.api
@@ -149,9 +149,7 @@ cmd const GLubyte* glGetString(GLenum param) {
   }
 
   ret := as!(const GLubyte*)(?)
-  if (ret != null) {
-    _ = as!string(as!(char*)(ret)) // Make sure we observe the string.
-  }
+  _ = as!string(as!(char*)(ret)) // Make sure we observe the string.
   return ret
 }
 
@@ -170,9 +168,7 @@ cmd const GLubyte* glGetStringi(GLenum name, GLuint index) {
   }
 
   ret := as!(const GLubyte*)(?)
-  if ret != null {
-    _ = as!string(as!(char*)(ret)) // Make sure we observe the string.
-  }
+  _ = as!string(as!(char*)(ret)) // Make sure we observe the string.
   return ret
 }