--- a/gapis/api/gles/gles_pb/BUILD.bazel
+++ b/gapis/api/gles/gles_pb/BUILD.bazel
@@ -34,7 +34,14 @@ go_library(
 )
 
 proto_library(
-    name = "gles_pb_proto",
+    name = "extras_proto",
+    srcs = ["extras.proto"],
+    visibility = ["//visibility:public"],
+    deps = ["//gapis/memory/memory_pb:memory_pb_proto"],  # keep
+)
+
+proto_library(
+    name = "api_pb_proto",
     srcs = [
         "extras.proto",
         ":api_proto",  # keep
@@ -44,15 +51,15 @@ proto_library(
 )
 
 cc_proto_library(
-    name = "gles_pb_cc_proto",
+    name = "extras_cc_proto",
     visibility = ["//visibility:public"],
-    deps = [":gles_pb_proto"],
+    deps = [":extras_proto"],
 )
 
 go_proto_library(
     name = "gles_pb_go_proto",
     importpath = "github.com/google/gapid/gapis/api/gles/gles_pb",
-    proto = ":gles_pb_proto",
+    proto = ":api_pb_proto",
     visibility = ["//visibility:public"],
     deps = ["//gapis/memory/memory_pb:go_default_library"],  # keep
 )