… gapii/cc/spy.cpp
… gapii/cc/to_proto.h
… gapil/compiler/BUILD.bazel
… gapis/api/templates/api_types.cpp.tmpl
… gapis/api/templates/api_types.h.tmpl