Optimized windows builds now work.